+++ src/widgets/label.rs
@@ -1,6 +1,6 @@
 //! A read-only text widget.
 
-use std::fmt::Write;
+use std::fmt::{Display, Write};
 
 use figures::units::{Px, UPx};
 use figures::{Point, Round, Size};
@@ -8,7 +8,7 @@ use kludgine::text::{MeasuredText, Text, TextOrigin};
 use kludgine::{CanRenderTo, Color, DrawableExt};
 
 use super::input::CowString;
-use crate::context::{GraphicsContext, LayoutContext, Trackable};
+use crate::context::{GraphicsContext, LayoutContext, Trackable, WidgetContext};
 use crate::styles::components::TextColor;
 use crate::value::{Dynamic, Generation, IntoReadOnly, ReadOnly, Value};
 use crate::widget::{Widget, WidgetInstance};
@@ -26,7 +26,7 @@ pub struct Label<T> {
 
 impl<T> Label<T>
 where
-    T: std::fmt::Debug + std::fmt::Display + Send + 'static,
+    T: std::fmt::Debug + DynamicDisplay + Send + 'static,
 {
     /// Returns a new label that displays `text`.
     pub fn new(text: impl IntoReadOnly<T>) -> Self {
@@ -54,7 +54,7 @@ where
                 context.apply_current_font_settings();
                 let measured = self.display.map(|text| {
                     self.displayed.clear();
-                    if let Err(err) = write!(&mut self.displayed, "{text}") {
+                    if let Err(err) = write!(&mut self.displayed, "{}", text.as_display(context)) {
                         tracing::error!("Error invoking Display: {err}");
                     }
                     context
@@ -75,7 +75,7 @@ where
 
 impl<T> Widget for Label<T>
 where
-    T: std::fmt::Debug + std::fmt::Display + Send + 'static,
+    T: std::fmt::Debug + DynamicDisplay + Send + 'static,
 {
     fn redraw(&mut self, context: &mut GraphicsContext<'_, '_, '_, '_>) {
         self.display.invalidate_when_changed(context);
@@ -132,3 +132,46 @@ impl_make_widget!(
     Value<String> => String,
     ReadOnly<String> => String
 );
+
+/// A context-aware [`Display`] implementation.
+///
+/// This trait is automatically implemented for all types that implement
+/// [`Display`].
+pub trait DynamicDisplay {
+    /// Format `self` with any needed information from `context`.
+    fn fmt(&self, context: &WidgetContext<'_>, f: &mut std::fmt::Formatter<'_>)
+        -> std::fmt::Result;
+
+    /// Returns a type that implements [`Display`].
+    fn as_display<'display, 'ctx>(
+        &'display self,
+        context: &'display WidgetContext<'ctx>,
+    ) -> DynamicDisplayer<'display, 'ctx>
+    where
+        Self: Sized,
+    {
+        DynamicDisplayer(self, context)
+    }
+}
+
+impl<T> DynamicDisplay for T
+where
+    T: Display,
+{
+    fn fmt(
+        &self,
+        _context: &WidgetContext<'_>,
+        f: &mut std::fmt::Formatter<'_>,
+    ) -> std::fmt::Result {
+        self.fmt(f)
+    }
+}
+
+/// A generic [`Display`] implementation for a [`DynamicDisplay`] implementor.
+pub struct DynamicDisplayer<'a, 'w>(&'a dyn DynamicDisplay, &'a WidgetContext<'w>);
+
+impl Display for DynamicDisplayer<'_, '_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.0.fmt(self.1, f)
+    }
+}